Helix Update - patch
Fun with Arrays Edition

Fixed:
- A case where no assigned game for a stream would result in a local
"game register" that didn't contain an expected key. Did a try/catch as
a simple fix to catch all future trouble as well.

- A case where getting the specialized followlists for the follow tab
would crash due to currentStreams being null as well as no follows
gathered. Added code that tries to retrieve follows before returning a
(now) struct with the lists and an isGood bool, in case of
currentFollows being null.